Retomando desenvolvimento

+++ PetAmigoApp/PetRescue/DeRuaViewController.m
@@ -37,7 +37,7 @@
     // Dispose of any resources that can be recreated.
 }
 
-- (int)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
+- (NSInteger)tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section {
     return cellIdentifier.count;
 }
